requirements.txt
django==3.2.3
aiogram==2.19
Pillow==9.0.1
asgiref==3.5.0
django-multiselectfield==0.1.12
gunicorn==20.1.0
dj-database-url==0.5.0
psycopg2==2.9.3